# openapi.yaml
openapi: 3.0.3
info:
  title: Invoice & Payment Service
  description: Minimal invoice and payment backend with idempotency, webhooks, and PSP integration.
  version: 0.1.0

servers:
  - url: http://localhost:8080
    description: Local development

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "API key in format: sk_test_<secret>"

  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "bad_request"
            message:
              type: string
              example: "At least one line item required"

    CreateCustomerRequest:
      type: object
      required: [email, name]
      properties:
        email:
          type: string
          format: email
        name:
          type: string

    CustomerResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        created_at:
          type: string
          format: date-time

    LineItemInput:
      type: object
      required: [description, quantity, unit_price_cents]
      properties:
        description:
          type: string
        quantity:
          type: integer
          minimum: 1
        unit_price_cents:
          type: integer
          minimum: 0
          description: "Price per unit in cents (integer, no floats)"

    CreateInvoiceRequest:
      type: object
      required: [customer_id, line_items]
      properties:
        customer_id:
          type: string
          format: uuid
        currency:
          type: string
          default: "usd"
        due_date:
          type: string
          format: date-time
        line_items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/LineItemInput'

    LineItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        quantity:
          type: integer
        unit_price_cents:
          type: integer
        amount_cents:
          type: integer
          description: "Server-computed: quantity * unit_price_cents"

    InvoiceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, paid, failed, void]
        amount_cents:
          type: integer
          description: "Server-computed total in cents"
        currency:
          type: string
        due_date:
          type: string
          format: date-time
          nullable: true
        line_items:
          type: array
          items:
            $ref: '#/components/schemas/LineItemResponse'
        created_at:
          type: string
          format: date-time

    PayInvoiceRequest:
      type: object
      required: [payment_token]
      properties:
        payment_token:
          type: string
          description: "Token from payment form (tok_success, tok_card_declined, etc.)"

    PaymentResponse:
      type: object
      properties:
        payment_id:
          type: string
          format: uuid
        invoice_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [succeeded, failed]
        amount_cents:
          type: integer
        failure_reason:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

paths:
  /api/customers:
    post:
      summary: Create a customer
      tags: [Customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '409':
          description: Customer with this email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List customers
      tags: [Customers]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CustomerResponse'

  /api/customers/{id}:
    get:
      summary: Get a customer by ID
      tags: [Customers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customer found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerResponse'
        '404':
          description: Customer not found

  /api/invoices:
    post:
      summary: Create an invoice
      description: |
        Creates an invoice with line items. The total is server-computed from
        (quantity * unit_price_cents) for each line item. Client-supplied totals are never accepted.
      tags: [Invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvoiceRequest'
      responses:
        '201':
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '400':
          description: Validation error (no line items, invalid customer, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: List invoices
      tags: [Invoices]
      parameters:
        - name: status
          in: query
          description: Filter by invoice status
          schema:
            type: string
            enum: [pending, paid, failed, void]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InvoiceResponse'

  /api/invoices/{id}:
    get:
      summary: Get an invoice by ID
      tags: [Invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Invoice found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceResponse'
        '404':
          description: Invoice not found

  /api/invoices/{id}/pay:
    post:
      summary: Pay an invoice
      description: |
        Initiates a payment attempt for the invoice. Requires an Idempotency-Key header.
        Uses SELECT FOR UPDATE to prevent concurrent double-charges.
        Calls the PSP with a 10-second timeout.
      tags: [Payments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
          description: "Unique key for idempotent retries. Reuse with different payload returns 422."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayInvoiceRequest'
      responses:
        '200':
          description: Payment attempt result (succeeded or failed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Invoice already paid or voided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Idempotency key reused with different payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'